+++ packages/reactive_phone_form_field/lib/src/validators/validation_message.dart
@@ -1,6 +1,4 @@
 class PhoneValidationMessage {
-  static const String required = 'phone.required';
   static const String valid = 'phone.valid';
   static const String validMobile = 'phone.validMobile';
-  static const String validFixedLine = 'phone.validFixedLine';
 }