+++ packages/reactive_phone_form_field/lib/src/validators/validator/valid_fixed_line_validator.dart
@@ -5,23 +5,44 @@ import 'package:reactive_phone_form_field/src/validators/validation_message.dart
 class ValidFixedLinePhoneValidator extends Validator<dynamic> {
   const ValidFixedLinePhoneValidator();
 
+  static const message = 'phone.validFixedLine';
+
   @override
   Map<String, dynamic>? validate(AbstractControl<dynamic> control) {
-    final error = <String, dynamic>{
-      PhoneValidationMessage.validFixedLine: true
-    };
+    final error = <String, dynamic>{ message: true };
 
-    if (control.value == null) {
-      return null;
-    } else if (control.value is PhoneNumber) {
-      PhoneNumber? valueCandidate = control.value as PhoneNumber;
+    final value = control.value;
 
-      if (PhoneValidator.validFixedLine().call(valueCandidate) == null) {
-        return null;
-      } else {
-        return error;
-      }
+    if(value is! PhoneNumber?) {
+      return error;
     }
+
+    if (
+        value.nsn.isNotEmpty &&
+        !valueCandidate.isValid(type: expectedType)) {
+      if (expectedType == PhoneNumberType.mobile) {
+        return errorText ??
+            PhoneFieldLocalization.of(context).invalidMobilePhoneNumber;
+      } else if (expectedType == PhoneNumberType.fixedLine) {
+        return errorText ??
+            PhoneFieldLocalization.of(context).invalidFixedLinePhoneNumber;
+      }
+      return errorText ??
+          PhoneFieldLocalization.of(context).invalidPhoneNumber;
+    }
+    return null;
+
+    // if (value == null) {
+    //   return null;
+    // } else if (control.value is PhoneNumber) {
+    //   PhoneNumber? valueCandidate = control.value as PhoneNumber;
+    //
+    //   if (PhoneValidator.validFixedLine().call(valueCandidate) == null) {
+    //     return null;
+    //   } else {
+    //     return error;
+    //   }
+    // }
 
     return null;
   }