+++ packages/reactive_dropdown_button2/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.0-beta.10
+
+* upgrade control value accessor to have ability to map against `items`
+
+## 0.0.4-beta.10
+
+* more styling options
+
+## 0.0.3-beta.10
+
+* wrapper for v3 beta.10
+* fix for disabled property
+
 ## 0.0.3-beta.6
 
 * wrapper for v3 beta.6