+++ packages/reactive_input_decorator/CHANGELOG.md
@@ -1,2 +1,4 @@
+## 0.0.2
+* ReactiveInputDecoratorStack
 ## 0.0.1
 * initial release